Optimization
- we were checking i === 0 || j === 0 in the innermost loop of damerau, and then skipping the one thing which made it different than standard levenshtein.
Instead of checking inside the loop, we can extract the first row and column and perform standard levenshtein on those.
This not only mildly quicker for the first row/column, but also removes the two comparisons from the inner loop.
In testing, this resulted in an approximately 50% increase in performance.