Stronger subtree filtering
Tracks the value and position of the best match over time rather than doing a linear search for the values when a candidate is found.
Minimal impact on performance, if anything slightly faster than previously.

Utilizes best match and trie depth to exit early for subtrees which are not capable of scoring high enough.
Reduces search times by as much as 50%.